remove old code & give dedicated directories for new types
