--- src/SocialOrganism.ts
@@ -1,7 +1,0 @@
-import { NeighbourhoodExpression } from "./Neighbourhood";
-import Perspective from "./perspectives/Perspective";
-
-export default class SocialOrganism {
-    inner: NeighbourhoodExpression
-    outer: Perspective
-}
