+++ src/social-organism/SocialOrganism.ts
@@ -1,0 +1,7 @@
+import { NeighbourhoodExpression } from "../neighbourhood/Neighbourhood";
+import Perspective from "../perspectives/Perspective";
+
+export default class SocialOrganism {
+    inner: NeighbourhoodExpression
+    outer: Perspective
+}
